Repackaging fixes for cmake-binary

+++ cmake-binary-feedstock/recipe/build.sh
@@ -1,7 +1,11 @@
 #!/bin/bash
 
-pushd cmake-bin/CMake.app/Contents
-  mkdir ${PREFIX}/cmake-bin
-  rm -rf bin/cmake-gui
-  cp -rf bin doc man share ${PREFIX}/cmake-bin
-popd
+if [[ $(uname) == Darwin ]]; then
+  pushd cmake-bin/CMake.app/Contents
+    mkdir ${PREFIX}/cmake-bin
+    rm -rf bin/cmake-gui
+    cp -rf bin doc man share ${PREFIX}/cmake-bin
+  popd
+else
+  cp -rf cmake-bin ${PREFIX}/
+fi
